Added some lab/instructor/location stuff.
Labs are not down to a science yet. Under tha category of labs falls
"Additional Lectures" and all that jazz. Which is kind of annoying,
because students don't get to pick which additional lecture they're in.
Essentially they're just another time for class. Now I'll need to work
on that.